+++ 1.21.11/common/src/main/java/com/iamkaf/liteminer/config/LiteminerClientConfig.java
@@ -7,6 +7,9 @@ public final class LiteminerClientConfig {
     public final ModConfigSpec.ConfigValue<Boolean> showHUD;
     public final ModConfigSpec.ConfigValue<Double> hud_scale;
 
+    public final ModConfigSpec.ConfigValue<LineColor> highlightForegroundLineColor;
+    public final ModConfigSpec.ConfigValue<LineColor> highlightSeeThroughLineColor;
+
     public LiteminerClientConfig(ModConfigSpec.Builder builder) {
         keyMode = builder.translation("liteminer.config.key_mode")
                 .comment(":)")
@@ -15,5 +18,13 @@ public final class LiteminerClientConfig {
         hud_scale = builder.translation("liteminer.config.hud_scale")
                 .comment(":)")
                 .defineInRange("hud_scale", 1d, 0.5d, 2d);
+
+        highlightForegroundLineColor = builder.translation("liteminer.config.highlight_foreground_line_color")
+                .comment(":)")
+                .defineEnum("highlight_foreground_line_color", LineColor.WHITE);
+
+        highlightSeeThroughLineColor = builder.translation("liteminer.config.highlight_see_through_line_color")
+                .comment(":)")
+                .defineEnum("highlight_see_through_line_color", LineColor.CYAN);
     }
 }